refactor(desktop): split workspacessidebar chrome into sub-files
