+++ apps/desktop/src/features/workspace/components/WorkspacesSidebar/lib.ts
@@ -1,0 +1,2 @@
+export const FOOTER_ICON_BTN =
+  'flex items-center justify-center rounded p-1.5 text-muted-foreground transition-colors hover:text-foreground hover:bg-muted/50' as const;
