+++ apps/desktop/src/features/workspace/components/WorkspacesSidebar/parts/SectionToggle.tsx
@@ -1,0 +1,21 @@
+import { ChevronDown, ChevronRight } from 'lucide-react';
+
+type Props = {
+  readonly expanded: boolean;
+  readonly label: string;
+  readonly onToggle: () => void;
+};
+
+export const SectionToggle = ({ expanded, label, onToggle }: Props) => {
+  return (
+    <button
+      type="button"
+      onClick={onToggle}
+      aria-expanded={expanded}
+      aria-label={`${expanded ? 'collapse' : 'expand'} ${label}`}
+      className="flex size-5 shrink-0 items-center justify-center rounded text-muted-foreground/50 transition-colors hover:bg-foreground/10 hover:text-foreground"
+    >
+      {expanded ? <ChevronDown size={12} aria-hidden /> : <ChevronRight size={12} aria-hidden />}
+    </button>
+  );
+};
